improved calendarImplem & tests
